.history/backend/requirements_20260209204257.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-dotenv
python-jose
passlib[bcrypt]
email-validator
python-multipart